PPF-382 - Move create coupon to helper function
