+++ e2e/tests/integrations/admin/create-coupon.test.ts
@@ -1,19 +1,11 @@
 import { expect, test } from "@playwright/test";
-import { faker } from "@faker-js/faker";
+import { createCoupon } from "./create-coupon.js";
 
 test.use({ storageState: "playwright/.auth/admin.json" });
 
 test("As an admin I can create a coupon)", async ({ page }) => {
-  await page.goto("/admin");
-  await page.getByRole("link", { name: "Coupons" }).click();
-  await page.getByRole("link", { name: "Create Coupon" }).click();
-  await page.waitForLoadState("networkidle");
-
-  const couponCode = faker.string.uuid();
-
-  await page.getByLabel("Coupon code").fill(couponCode);
-  await page.getByRole("button", { name: "Create Coupon" }).click();
-  await page.waitForLoadState("networkidle");
+ 
+  const { couponCode } = await createCoupon(page);
 
   await expect(
     page.locator("h1").getByText(`Coupon Details: ${couponCode}`)
